Set default-directory in ssh-tunnels--command
Prior to this commit, callers of ssh-tunnels--run, ssh-tunnels--kill, and
ssh-tunnels--check were responsible for binding default-directory to ensure
that SSH used the correct directory for control sockets.  This commit
relieves callers of this burden by changing ssh-tunnels--command to set
default-directory.

* ssh-tunnels.el (ssh-tunnels--command): Bind default-directory to
ssh-tunnels-temp-directory before running SSH.
(open-network-stream@run-ssh-tunnel): Delete binding of default-directory;
it is now superfluous.